LayoutTests/plugins/navigator-plugins.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description(
    "Check that navigator.plugins does not contain empty plugin objects"
);

var hasInvalidPlugin = false;
for (var i = 0; i < navigator.plugins.length; ++i) {
    var plugin = navigator.plugins[i];
    if (!plugin.description || !plugin.name || !plugin.filename) {
        hasInvalidPlugin = true;
        break;
    }
}
shouldBeFalse("hasInvalidPlugin");

successfullyParsed = true;
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
