third_party/blink/web_tests/platform/generic/plugins/navigator-plugins-expected.txt

Check that navigator.plugins does not contain empty plugin objects

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS hasInvalidPlugin is false
PASS successfullyParsed is true

TEST COMPLETE
